structura
дода папку, винесла деякі функції
