+++ src/index.js
@@ -1,9 +1,9 @@
 import './js/cardModal';
-import { refs, renderGallery } from './js/renderGallery';
+import { refs, renderGallery } from './js/renderHomeGallery';
 import './js/footer_modal';
-import './js/renderGallery';
-import './js/spinner';
-import './js/userLists';
+import './js/renderHomeGallery';
+import './js/templates/spinner';
+import './js/renderLibraryLists';
 
 import * as Snow from './js/efects/snow';
   const snow = new Snow.default({
