+++ src/js/templates/refs.js
@@ -1,0 +1,8 @@
+export const refs = {
+  searchForm: document.querySelector('#search-form'),
+  searchInput: document.querySelector('.hero__form-input'),
+  genreField: document.querySelector('.genre'),
+  gallery: document.querySelector('.gallery__list'),
+  homePage: document.querySelector('#home-page'),
+  searchIcon: document.querySelector('.hero__search-icon'),
+};
