+++ src/js/templates/spinner.js
